changed entity id from long to int

+++ src/framework/ecs/ObjectManager.cpp
@@ -16,7 +16,7 @@ namespace HGE {
         start = 0;
     }
 
-    Object* ObjectManager::GetObject(long uid) {
+    Object* ObjectManager::GetObject(int uid) {
         auto it = std::find_if(mObjects.begin(), mObjects.end(), [&uid] (const std::unique_ptr<Object> & pObject) { 
             return pObject->getId() == uid;
             });
@@ -24,7 +24,7 @@ namespace HGE {
             return it->get();
     }
 
-    void ObjectManager::DestroyObject(long uid) {
+    void ObjectManager::DestroyObject(int uid) {
             auto it = std::find_if(mObjects.begin(), mObjects.end(), [&uid] (const std::unique_ptr<Object> & pObject) { 
             return pObject->getId() == uid;
             });
